requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
requests
psycopg2-binary
python-dotenv
redis